Move getBehindBall to active tactics
