+++ src/stp/tactics/active/GetBehindBallInDirection.cpp
@@ -2,7 +2,7 @@
 // Created by jordi on 06-04-20.
 // Rebuilt by alexander on 21-12-21
 //
-#include "stp/tactics/GetBehindBallInDirection.h"
+#include "stp/tactics/active/GetBehindBallInDirection.h"
 
 #include <roboteam_utils/LineSegment.h>
